+++ lisa-kernel/src/main/scala/lisa/kernel/proof/SequentCalculus.scala
@@ -191,7 +191,7 @@ object SequentCalculus {
 
   /**
    * <pre>
-   *  Γ |- φ→ψ, Δ    Σ |- ψ→φ, Π
+   *  Γ |- a→ψ, Δ    Σ |- ψ→φ, Π
    * ----------------------------
    *      Γ, Σ |- φ↔ψ, Π, Δ
    * </pre>
